pyproject.toml
[tool.poetry]
name = "gstsol"
version = "0.1.0"
description = ""
authors = ["Yoyo <jack456054@hotmail.com>"]

[tool.poetry.dependencies]
python = "3.8.13"
requests = "^2.27.1"
black = "^22.3.0"
plotly = "^5.8.0"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"
Flask = "^2.1.2"
line-bot-sdk = "^2.2.1"
kaleido = "0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
